spatial-environment-agent/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
PyYAML==6.0.2
httpx==0.28.1
